Fixed GMT timezone bug

+++ app/src/main/java/com/plavsic/skytrace/features/weather/model/WeatherInfo.kt
@@ -1,10 +1,10 @@
 package com.plavsic.skytrace.features.weather.model
 
 data class WeatherInfo(
-    val temp: Double,
-    val feels_like: Double,
-    val temp_min: Double,
-    val temp_max: Double,
+    val temp: Int,
+    val feels_like: Int,
+    val temp_min: Int,
+    val temp_max: Int,
     val pressure: Int,
     val humidity: Int
 )
